% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.marketplacemetering_operations.R
\name{batch_meter_usage}
\alias{batch_meter_usage}
\title{BatchMeterUsage is called from a SaaS application listed on the AWS Marketplace to post metering records for a set of customers}
\usage{
batch_meter_usage(
  UsageRecords = list(
    list(
      Timestamp = as.POSIXct("2015-01-01"),
      CustomerIdentifier = "string",
      Dimension = "string",
      Quantity = 123
    )
  ),
  ProductCode = "string"
)
}
\arguments{
\item{UsageRecords}{[required] The set of UsageRecords to submit. BatchMeterUsage accepts up to 25 UsageRecords at a time.}

\item{ProductCode}{[required] Product code is used to uniquely identify a product in AWS Marketplace. The product code should be the same as the one used during the publishing of a new product.}
}
\description{
BatchMeterUsage is called from a SaaS application listed on the AWS Marketplace to post metering records for a set of customers.
}
\details{
For identical requests, the API is idempotent; requests can be retried with the same records or a subset of the input records.

Every request to BatchMeterUsage is for one product. If you need to meter usage for multiple products, you must make multiple calls to BatchMeterUsage.

BatchMeterUsage can process up to 25 UsageRecords at a time.
}
\examples{

}
