% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.marketplacemetering_package.R
\docType{package}
\name{paws.marketplacemetering-package}
\alias{paws.marketplacemetering}
\alias{paws.marketplacemetering-package}
\title{paws.marketplacemetering: AWSMarketplace Metering}
\description{
This reference provides descriptions of the
low-level AWS Marketplace Metering Service API.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
