% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.marketplacemetering_operations.R
\name{resolve_customer}
\alias{resolve_customer}
\title{ResolveCustomer is called by a SaaS application during the registration process}
\usage{
resolve_customer(RegistrationToken)
}
\arguments{
\item{RegistrationToken}{[required] When a buyer visits your website during the registration process, the buyer submits a registration token through the browser. The registration token is resolved to obtain a CustomerIdentifier and product code.}
}
\description{
ResolveCustomer is called by a SaaS application during the registration process. When a buyer visits your website during the registration process, the buyer submits a registration token through their browser. The registration token is resolved through this API to obtain a CustomerIdentifier and product code.
}
\section{Accepted Parameters}{
\preformatted{resolve_customer(
  RegistrationToken = "string"
)
}
}

